chore: moved service-offering type to IgxComplianceClient
